+++ packages/conc/changelog.md
@@ -6,8 +6,9 @@
 * Add `Semaphore`, abstracting `QSem` or `TSem`.
 * Add `timeoutStop`, a variant of `timeout` that calls `Polysemy.Resume.stop`.
 * Add `Lock`, a mutex effect.
-* Add `withAsyncTrigger`, a variant of `withAsync` that requires the async action to signal the sync action to start.
+* Add `withAsyncGated`, a variant of `withAsync` that requires the async action to signal the sync action to start.
 * Add `Gate`, an synchronization point effect.
+* Add variants of `subscribeWhile` and `subscribeLoop` that use `Gate` for synchronization and run in a new thread.
 
 # 0.9.0.0